requirements.txt
asgiref==3.3.4
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
colorama==0.4.4
decorator==5.0.9
Django==3.2.4
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
Pygments==2.9.0
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5

djangorestframework~=3.12.4
